chore(monorepo): `rehype-shiki`
