+++ apps/site/components/MDX/CodeBox/index.tsx
@@ -1,7 +1,7 @@
+import { getLanguageDisplayName } from '@node-core/rehype-shiki';
 import type { FC, PropsWithChildren } from 'react';
 
 import CodeBox from '#site/components/Common/CodeBox';
-import { getLanguageDisplayName } from '#site/util/getLanguageDisplayName';
 
 type CodeBoxProps = { className?: string; showCopyButton?: string };
