+++ apps/site/next.mdx.plugins.mjs
@@ -1,12 +1,11 @@
 'use strict';
 
+import rehypeShikiji from '@node-core/rehype-shiki';
 import remarkHeadings from '@vcarl/remark-headings';
 import rehypeAutolinkHeadings from 'rehype-autolink-headings';
 import rehypeSlug from 'rehype-slug';
 import remarkGfm from 'remark-gfm';
 import readingTime from 'remark-reading-time';
-
-import rehypeShikiji from './next.mdx.shiki.mjs';
 
 /**
  * Provides all our Rehype Plugins that are used within MDX
