+++ packages/rehype-shiki/src/index.mjs
@@ -1,0 +1,5 @@
+import { rehypeShikiji } from './plugin.mjs';
+export * from './highlighter.mjs';
+export * from './languages.mjs';
+
+export default rehypeShikiji;
